# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Freerock Solar & Energy API",
    "version": "1.0.0",
    "description": "Production PWA for Zimbabwe's trusted solar partner — instant quotes, Paynow payments, training enrollment, admin dashboard.\n\nBase URL: `https://freerock.co.zw` (production) or `http://localhost:3000` (development).\n\nAll endpoints that accept JSON use `Content-Type: application/json`. Protected admin routes require a valid Supabase session cookie (set via `/api/auth/otp` or admin login)."
  },
  "servers": [
    { "url": "https://freerock.co.zw", "description": "Production" },
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
...
  "paths": {
    "/api/quote": {
      "post": {
        "tags": ["Quotes"],
        "summary": "Create a solar quote",
        "description": "Calculates pricing server-side using the shared pricing engine (pricing.ts). Returns a canonical quote object regardless of Supabase persistence state.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["fullName", "phone", "province", "propertyType", "roofType", "serviceId"],
                "properties": {
                  "fullName": { "type": "string", "example": "Tatenda Masuku" },
                  "email": { "type": "string", "format": "email", "description": "Optional" },
                  "phone": { "type": "string", "example": "+263 77 123 4567" },
                  "province": { "type": "string", "example": "Harare" },
                  "propertyType": { "type": "string", "enum": ["residential", "commercial", "industrial"] },
                  "roofType": { "type": "string", "enum": ["tile", "corrugated", "flat", "ground"] },
                  "serviceId": { "type": "string", "example": "5.5kva" },
                  "payAfterInstall": { "type": "boolean", "default": true },
                  "addons": {
                    "type": "object",
                    "properties": {
                      "tvBundle": { "type": "boolean" },
                      "extraPanels": { "type": "integer", "minimum": 0 }
                    }
                  },
                  "notes": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Quote created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "quote": { "$ref": "#/components/schemas/Quote" }
                  }
                }
              }
            }
          },
          "400": { "description": "Validation failed" },
          "429": { "description": "Rate limited (10 req/h per IP)" }
        }
      }
    },
    "/api/quote/{id}": {
      "get": {
        "tags": ["Quotes"],
        "summary": "Get a quote by ID",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" }, "example": "FRQ-202607-0001" }
        ],
        "responses": {
          "200": {
            "description": "Quote found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "quote": { "$ref": "#/components/schemas/Quote" }
                  }
                }
              }
            }
          },
          "404": { "description": "Quote not found" }
        }
      },
      "patch": {
        "tags": ["Quotes"],
        "summary": "Update quote status",
        "description": "Used by admin leads page and payment confirmation to mark payment_status or quote_status.",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "paymentStatus": { "type": "string", "enum": ["pending", "deposit_paid", "fully_paid", "refunded"] },
                  "quoteStatus": { "type": "string", "enum": ["pending", "approved", "scheduled", "in_progress", "completed", "cancelled"] },
                  "paymentMethod": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated quote" },
          "400": { "description": "Validation failed" }
        }
      },
      "delete": {
        "tags": ["Quotes"],
        "summary": "Delete a quote",
        "parameters": [
          { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "id": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/rate": {
      "get": {
        "tags": ["System"],
        "summary": "Get current USD/ZIG exchange rate",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pair": { "type": "string", "example": "USD/ZIG" },
                    "rate": { "type": "number" },
                    "source": { "type": "string" },
                    "lastUpdated": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["System"],
        "summary": "Override exchange rate (admin)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["rate"],
                "properties": {
                  "rate": { "type": "number", "exclusiveMinimum": 0 }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean" },
                    "rate": { "type": "number" },
                    "updatedAt": { "type": "string", "format": "date-time" }
                  }
                }
              }
            }
          },
          "400": { "description": "Invalid rate value" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Customer": {
        "type": "object",
        "properties": {
          "fullName": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "phone": { "type": "string" },
          "province": { "type": "string" },
          "propertyType": { "type": "string" },
          "roofType": { "type": "string" }
        }
      },
      "QuotePricing": {
        "type": "object",
        "properties": {
          "packagePrice": { "type": "number" },
          "tvBundlePrice": { "type": "number" },
          "installationPrice": { "type": "number" },
          "extraPanelPrice": { "type": "number" },
          "auditPrice": { "type": "number" },
          "subtotal": { "type": "number" },
          "depositRate": { "type": "number" },
          "depositAmount": { "type": "number" },
          "balance": { "type": "number" }
        }
      },
      "Quote": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "example": "FRQ-202607-0042" },
          "customer": { "$ref": "#/components/schemas/Customer" },
          "serviceId": { "type": "string" },
          "serviceName": { "type": "string" },
          "pricing": { "$ref": "#/components/schemas/QuotePricing" },
          "addons": {
            "type": "object",
            "properties": {
              "tvBundle": { "type": "boolean" }
            }
          },
          "payAfterInstall": { "type": "boolean" },
          "status": { "type": "string", "enum": ["pending", "approved", "scheduled", "in_progress", "completed", "cancelled"] },
          "paymentStatus": { "type": "string", "enum": ["pending", "deposit_paid", "fully_paid", "refunded"] },
          "notes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  }
}